requirements.txt
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.13.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
numpy==2.3.5
openai==2.9.0
packaging==25.0
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1